Viewprotocols extration: some code has been moved away to be lighter
New window: SearchRun to search for a run in the graph. Using a new SearchBaseWindow
Searchprotocol reaftored to use the new SearchBase window
