+++ CHANGES.txt
@@ -1,3 +1,8 @@
+V3.0.20
+-------
+ - New "Search protocol in graph" (Control + l) window to locate a box in the graph
+ - At loading time, selected protocol is "focused" (window scrolls to it)
+
 V3.0.19
 -------
  - Empty sections/subsection (without protocols) in the left tree will not be shown
