+++ pyworkflow/gui/project/viewprotocols.py
@@ -22,10 +22,6 @@
 # *  e-mail address 'scipion@cnb.csic.es'
 # *
 # **************************************************************************
-from configparser import ConfigParser
-
-from pyworkflow.project import MenuConfig
-from pyworkflow import Config, TK
 
 INIT_REFRESH_SECONDS = 5
 
@@ -42,44 +38,24 @@ import tkinter as tk
 import tkinter.ttk as ttk
 import datetime as dt
 
-import pyworkflow.object as pwobj
+from pyworkflow import Config, TK
 import pyworkflow.utils as pwutils
 import pyworkflow.protocol as pwprot
-import pyworkflow.gui as pwgui
 from pyworkflow.viewer import DESKTOP_TKINTER, ProtocolViewer
-from pyworkflow.utils.properties import Message, Icon, Color, KEYSYM
-from pyworkflow.gui.project.utils import getStatusColorFromNode
-from pyworkflow.gui.form import FormWindow
+from pyworkflow.utils.properties import Color, KEYSYM
 from pyworkflow.webservices import WorkflowRepository
+
+import pyworkflow.gui as pwgui
+from pyworkflow.gui.form import FormWindow
+from pyworkflow.gui.project.utils import getStatusColorFromNode, inspectObj
+from pyworkflow.gui.project.constants import *
+from pyworkflow.gui.project.searchprotocol import SearchProtocolWindow, ProtocolTreeProvider
+from pyworkflow.gui.project.steps import StepsWindow
+from pyworkflow.gui.project.viewprotocols_extra import RunIOTreeProvider, ProtocolTreeConfig
+from pyworkflow.gui.project.searchrun import RunsTreeProvider, SearchRunWindow
 
 DEFAULT_BOX_COLOR = '#f8f8f8'
 
-ACTION_EDIT = Message.LABEL_EDIT
-ACTION_RENAME = Message.LABEL_RENAME
-ACTION_SELECT_FROM = Message.LABEL_SELECT_FROM
-ACTION_SELECT_TO = Message.LABEL_SELECT_TO
-ACTION_COPY = Message.LABEL_COPY
-ACTION_DELETE = Message.LABEL_DELETE
-ACTION_REFRESH = Message.LABEL_REFRESH
-ACTION_STEPS = Message.LABEL_STEPS
-ACTION_BROWSE = Message.LABEL_BROWSE
-ACTION_DB = Message.LABEL_DB
-ACTION_TREE = Message.LABEL_TREE
-ACTION_LIST = Message.LABEL_LIST
-ACTION_STOP = Message.LABEL_STOP
-ACTION_DEFAULT = Message.LABEL_DEFAULT
-ACTION_CONTINUE = Message.LABEL_CONTINUE
-ACTION_RESULTS = Message.LABEL_ANALYZE
-ACTION_EXPORT = Message.LABEL_EXPORT
-ACTION_EXPORT_UPLOAD = Message.LABEL_EXPORT_UPLOAD
-ACTION_SWITCH_VIEW = 'Switch_View'
-ACTION_COLLAPSE = 'Collapse'
-ACTION_EXPAND = 'Expand'
-ACTION_LABELS = 'Labels'
-ACTION_RESTART_WORKFLOW = Message.LABEL_RESTART_WORKFLOW
-ACTION_CONTINUE_WORKFLOW = Message.LABEL_CONTINUE_WORKFLOW
-ACTION_STOP_WORKFLOW = Message.LABEL_STOP_WORKFLOW
-ACTION_RESET_WORKFLOW = Message.LABEL_RESET_WORKFLOW
 
 RUNS_TREE = Icon.RUNS_TREE
 RUNS_LIST = Icon.RUNS_LIST
@@ -87,531 +63,6 @@ RUNS_LIST = Icon.RUNS_LIST
 VIEW_LIST = 0
 VIEW_TREE = 1
 VIEW_TREE_SMALL = 2
-
-ActionIcons = {
-    ACTION_EDIT: Icon.ACTION_EDIT,
-    ACTION_SELECT_FROM: Icon.ACTION_SELECT_FROM,
-    ACTION_SELECT_TO: Icon.ACTION_SELECT_TO,
-    ACTION_COPY: Icon.ACTION_COPY,
-    ACTION_DELETE: Icon.ACTION_DELETE,
-    ACTION_REFRESH: Icon.ACTION_REFRESH,
-    ACTION_RENAME: Icon.ACTION_RENAME,
-    ACTION_STEPS: Icon.ACTION_STEPS,
-    ACTION_BROWSE: Icon.ACTION_BROWSE,
-    ACTION_DB: Icon.ACTION_DB,
-    ACTION_TREE: None,  # should be set
-    ACTION_LIST: Icon.ACTION_LIST,
-    ACTION_STOP: Icon.ACTION_STOP,
-    ACTION_CONTINUE: Icon.ACTION_CONTINUE,
-    ACTION_RESULTS: Icon.ACTION_RESULTS,
-    ACTION_EXPORT: Icon.ACTION_EXPORT,
-    ACTION_EXPORT_UPLOAD: Icon.ACTION_EXPORT_UPLOAD,
-    ACTION_COLLAPSE: 'fa-minus-square.gif',
-    ACTION_EXPAND: 'fa-plus-square.gif',
-    ACTION_LABELS: Icon.TAGS,
-    ACTION_RESTART_WORKFLOW: Icon.ACTION_EXECUTE,
-    ACTION_CONTINUE_WORKFLOW: Icon.ACTION_CONTINUE,
-    ACTION_STOP_WORKFLOW: Icon.ACTION_STOP_WORKFLOW,
-    ACTION_RESET_WORKFLOW: Icon.ACTION_REFRESH
-}
-
-
-class RunsTreeProvider(pwgui.tree.ProjectRunsTreeProvider):
-    """Provide runs info to populate tree"""
-
-    def __init__(self, project, actionFunc):
-        pwgui.tree.ProjectRunsTreeProvider.__init__(self, project)
-        self.actionFunc = actionFunc
-        self._selection = project.getSettings().runSelection
-
-    def getActionsFromSelection(self):
-        """ Return the list of options available for selection. """
-        n = len(self._selection)
-        single = n == 1
-        if n:
-            prot = self.project.getProtocol(self._selection[0])
-            status = prot.getStatus()
-            nodeInfo = self.project.getSettings().getNodeById(prot.getObjId())
-            expanded = nodeInfo.isExpanded() if nodeInfo else True
-        else:
-            status = None
-
-        stoppable = status in [pwprot.STATUS_RUNNING, pwprot.STATUS_SCHEDULED, 
-                               pwprot.STATUS_LAUNCHED]
-
-        return [(ACTION_EDIT, single and status and expanded),
-                (ACTION_RENAME, single and status and expanded),
-                (ACTION_COPY, status and expanded),
-                (ACTION_DELETE, status != pwprot.STATUS_RUNNING and status and expanded),
-                (ACTION_STEPS, single and Config.debugOn() and status and expanded),
-                (ACTION_BROWSE, single and status and expanded),
-                (ACTION_DB, single and Config.debugOn() and status and expanded),
-                (ACTION_STOP, stoppable and single),
-                (ACTION_EXPORT, not single),
-                (ACTION_EXPORT_UPLOAD, not single),
-                (ACTION_COLLAPSE, single and status and expanded),
-                (ACTION_EXPAND, single and status and not expanded),
-                (ACTION_LABELS, True),
-                (ACTION_SELECT_FROM, True),
-                (ACTION_SELECT_TO, True),
-                (ACTION_RESTART_WORKFLOW, single),
-                (ACTION_CONTINUE_WORKFLOW, single),
-                (ACTION_STOP_WORKFLOW, single),
-                (ACTION_RESET_WORKFLOW, single)
-                ]
-
-    def getObjectActions(self, obj):
-
-        def addAction(actionLabel):
-            if actionLabel:
-                text = actionLabel
-                action = actionLabel
-                actionLabel = (text, lambda: self.actionFunc(action),
-                               ActionIcons.get(action, None))
-            return actionLabel
-
-        actions = [addAction(a)
-                   for a, cond in self.getActionsFromSelection() if cond]
-
-        if hasattr(obj, 'getActions'):
-            for text, action in obj.getActions():
-                actions.append((text, action, None))
-
-        return actions
-
-
-class ProtocolTreeProvider(pwgui.tree.ObjectTreeProvider):
-    """Create the tree elements for a Protocol run"""
-
-    def __init__(self, protocol):
-        self.protocol = protocol
-        # This list is create to group the protocol parameters
-        # in the tree display
-        self.status = pwobj.List(objName='_status')
-        self.params = pwobj.List(objName='_params')
-        self.statusList = ['status', 'initTime', 'endTime', 'error',
-                           'interactive', 'mode']
-
-        objList = [] if protocol is None else [protocol]
-        pwgui.tree.ObjectTreeProvider.__init__(self, objList)
-
-
-class StepsTreeProvider(pwgui.tree.TreeProvider):
-    """Create the tree elements for a Protocol run"""
-
-    def __init__(self, stepsList):
-        for i, s in enumerate(stepsList):
-            if not s._index:
-                s._index = i + 1
-
-        self._stepsList = stepsList
-        self.getColumns = lambda: [('Index', 50), ('Step', 200), ('Status', 150),
-                                   ('Time', 150), ('Class', 100)]
-        self._parentDict = {}
-
-    def getObjects(self):
-        return self._stepsList
-
-    @staticmethod
-    def getObjectInfo(obj):
-        info = {'key': obj._index,
-                'values': (str(obj), obj.getStatus(), pwutils.prettyDelta(obj.getElapsedTime()),
-                           obj.getClassName())}
-        return info
-
-    @staticmethod
-    def getObjectPreview(obj):
-
-        args = json.loads(obj.argsStr.get())
-        msg = "*Prerequisites*: %s \n" % str(obj._prerequisites)
-        msg += "*Arguments*: " + '\n  '.join([str(a) for a in args])
-        if hasattr(obj, 'resultFiles'):
-            results = json.loads(obj.resultFiles.get())
-            if len(results):
-                msg += "\n*Result files:* " + '\n  '.join(results)
-
-        return None, msg
-
-
-class StepsWindow(pwgui.browser.BrowserWindow):
-    def __init__(self, title, parentWindow, protocol, **args):
-        self._protocol = protocol
-        provider = StepsTreeProvider(protocol.loadSteps())
-        pwgui.browser.BrowserWindow.__init__(self, title, parentWindow,
-                                             weight=False, **args)
-        # Create buttons toolbar
-        self.root.columnconfigure(0, weight=1)
-        self.root.rowconfigure(1, weight=1)
-
-        self.fillToolBar()
-
-        # Create and set browser
-        browser = pwgui.browser.ObjectBrowser(self.root, provider,
-                                              showPreviewTop=False)
-        self.setBrowser(browser, row=1, column=0)
-
-    def fillToolBar(self):
-        # Tool bar
-        toolbar = tk.Frame(self.root)
-        toolbar.grid(row=0, column=0, sticky='nw', padx=5, pady=5)
-
-        # Tree button
-        btn = tk.Label(toolbar, text="Tree",
-                       image=self.getImage(Icon.RUNS_TREE),
-                       compound=tk.LEFT, cursor='hand2')
-        btn.bind(TK.LEFT_CLICK, self._showTree)
-        btn.grid(row=0, column=0, sticky='nw')
-
-        # Reset status
-        btn = tk.Label(toolbar, text="Reset",
-                       image=self.getImage(Icon.ACTION_REFRESH),
-                       compound=tk.LEFT, cursor='hand2')
-        btn.bind('<Button-1>', self._resetStep)
-        btn.grid(row=0, column=1, sticky='nw')
-
-    def _resetStep(self, e=None):
-
-        item = self.browser._lastSelected
-        if item is not None:
-            objId = item.getObjId()
-            self._protocol._updateSteps(lambda step: step.setStatus(pwprot.STATUS_NEW), where="id='%s'" % objId)
-            item.setStatus(pwprot.STATUS_NEW)
-            self.browser.tree.update()
-    # noinspection PyUnusedLocal
-    def _showTree(self, e=None):
-        g = self._protocol.getStepsGraph()
-        w = pwgui.Window("Protocol steps", self, minsize=(800, 600))
-        root = w.root
-        canvas = pwgui.Canvas(root, width=600, height=500,
-                              tooltipCallback=self._stepTooltip,)
-        canvas.grid(row=0, column=0, sticky='nsew')
-        canvas.drawGraph(g, pwgui.LevelTreeLayout())
-        w.show()
-
-    def _stepTooltip(self, tw, item):
-        """ Create the contents of the tooltip to be displayed
-        for the given step.
-        Params:
-            tw: a tk.TopLevel instance (ToolTipWindow)
-            item: the selected step.
-        """
-
-        if not hasattr(item.node, 'step'):
-            return
-
-        step = item.node.step
-
-        tm = str(step.funcName)
-
-        if not hasattr(tw, 'tooltipText'):
-            frame = tk.Frame(tw)
-            frame.grid(row=0, column=0)
-            tw.tooltipText = pwgui.dialog.createMessageBody(
-                frame, tm, None, textPad=0, textBg=Color.LIGHT_GREY_COLOR_2)
-            tw.tooltipText.config(bd=1, relief=tk.RAISED)
-        else:
-            pwgui.dialog.fillMessageText(tw.tooltipText, tm)
-
-
-class SearchProtocolWindow(pwgui.Window):
-    def __init__(self, parentWindow, **kwargs):
-        pwgui.Window.__init__(self, title="Search for a protocol",
-                              masterWindow=parentWindow)
-        content = tk.Frame(self.root, bg='white')
-        self._createContent(content)
-        content.grid(row=0, column=0, sticky='news')
-        content.columnconfigure(0, weight=1)
-        content.rowconfigure(1, weight=1)
-
-    def _createContent(self, content):
-        self._createSearchBox(content)
-        self._createResultsBox(content)
-
-    def _createSearchBox(self, content):
-        """ Create the Frame with Search widgets """
-        frame = tk.Frame(content, bg='white')
-
-        label = tk.Label(frame, text="Search", bg='white')
-        label.grid(row=0, column=0, sticky='nw')
-        self._searchVar = tk.StringVar()
-        entry = tk.Entry(frame, bg='white', textvariable=self._searchVar)
-        entry.bind(TK.RETURN, self._onSearchClick)
-        entry.bind(TK.ENTER, self._onSearchClick)
-        entry.focus_set()
-        entry.grid(row=0, column=1, sticky='nw')
-        btn = pwgui.widgets.IconButton(frame, "Search",
-                                       imagePath=Icon.ACTION_SEARCH,
-                                       command=self._onSearchClick)
-        btn.grid(row=0, column=2, sticky='nw')
-
-        frame.grid(row=0, column=0, sticky='new', padx=5, pady=(10, 5))
-
-    def _createResultsBox(self, content):
-        frame = tk.Frame(content, bg=Color.LIGHT_GREY_COLOR, padx=5, pady=5)
-        pwgui.configureWeigths(frame)
-        self._resultsTree = self.master.getViewWidget()._createProtocolsTree(
-            frame, show=None, columns=("protocol", "streaming", "installed", "help", "score"))
-        self._configureTreeColumns()
-        self._resultsTree.grid(row=0, column=0, sticky='news')
-        frame.grid(row=1, column=0, sticky='news', padx=5, pady=5)
-
-    def _configureTreeColumns(self):
-        self._resultsTree.column('#0', width=50, minwidth=50, stretch=tk.NO)
-        self._resultsTree.column('protocol', width=300, stretch=tk.FALSE)
-        self._resultsTree.column('streaming', width=100, stretch=tk.FALSE)
-        self._resultsTree.column('installed', width=110, stretch=tk.FALSE)
-        self._resultsTree.column('help', minwidth=300, stretch=tk.YES)
-        self._resultsTree.column('score', width=50, stretch=tk.FALSE)
-
-        self._resultsTree.heading('#0', text='Status')
-        self._resultsTree.heading('protocol', text='Protocol', command=lambda: self._resultsTree.sortByColumn("protocol", False))
-        self._resultsTree.heading('streaming', text='Streamified', command=lambda: self._resultsTree.sortByColumn("streaming", False))
-        self._resultsTree.heading('installed', text='Installation', command=lambda: self._resultsTree.sortByColumn("installed", False))
-        self._resultsTree.heading('help', text='Help', command=lambda: self._resultsTree.sortByColumn("help", False))
-        self._resultsTree.heading('score', text='Score', command=lambda: self._resultsTree.sortByColumn("score", False, casting=int))
-
-    def _onSearchClick(self, e=None):
-        self._resultsTree.clear()
-        keyword = self._searchVar.get().lower().strip()
-        emProtocolsDict = Config.getDomain().getProtocols()
-        protList = []
-
-        def addSearchWeight(line2Search, searchtext):
-            # Adds a weight value for the search
-            weight = 0
-
-            # prioritize findings in label
-            if searchtext in line2Search[1]:
-                weight += 10
-
-            for value in line2Search[2:]:
-                weight += 5 if searchtext in value else 0
-
-            if " " in searchtext:
-                for word in searchtext.split():
-                    if word in line2Search[1]:
-                        weight += 3
-
-                    for value in line2Search[2:]:
-                        weight += 1 if word in value else 0
-
-            return line2Search + (weight,)
-
-        for key, prot in emProtocolsDict.items():
-            if ProtocolTreeConfig.isAFinalProtocol(prot, key):
-                label = prot.getClassLabel().lower()
-                line = (key, label,
-                        "installed" if prot.isInstalled() else "missing installation",
-                        prot.getHelpText().strip().replace('\r', '').replace('\n', '').lower(),
-                        "streamified" if prot.worksInStreaming() else "static",
-                        "beta" if prot.isBeta() else "",
-                        "new" if prot.isNew() else "")
-
-                line = addSearchWeight(line, keyword)
-                # something was found: weight > 0
-                if line[7] != 0:
-                    protList.append(line)
-
-        # Sort by weight
-        protList.sort(reverse=True, key=lambda x: x[7])
-
-        for key, label, installed, help, streamified, beta, new, weight in protList:
-            tag = ProtocolTreeConfig.getProtocolTag(installed == 'installed',
-                                                    beta == 'beta',
-                                                    new == 'new')
-            self._resultsTree.insert(
-                '', 'end', key, text="", tags=tag,
-                values=(label, streamified, installed, help, weight))
-
-
-class RunIOTreeProvider(pwgui.tree.TreeProvider):
-    """Create the tree elements from a Protocol Run input/output childs"""
-
-    def __init__(self, parent, protocol, mapper):
-        # TreeProvider.__init__(self)
-        self.parent = parent
-        self.protocol = protocol
-        self.mapper = mapper
-
-    @staticmethod
-    def getColumns():
-        return [('Attribute', 200), ('Info', 100)]
-
-    def getObjects(self):
-        objs = []
-        if self.protocol is not None:
-            # Store a dict with input parents (input, PointerList)
-            self.inputParentDict = OrderedDict()
-            inputs = []
-            inputObj = pwobj.String(Message.LABEL_INPUT)
-            inputObj._icon = Icon.ACTION_IN
-            self.inputParentDict['_input'] = inputObj
-            inputParents = [inputObj]
-
-            for key, attr in self.protocol.iterInputAttributes():
-                attr._parentKey = key
-                # Repeated keys means there are inside a pointerList
-                # since the same key is yielded for all items inside
-                # so update the parent dict with a new object
-                if key in self.inputParentDict:
-                    if self.inputParentDict[key] == inputObj:
-                        parentObj = pwobj.String(key)
-                        parentObj._icon = Icon.ACTION_IN
-                        parentObj._parentKey = '_input'
-                        inputParents.append(parentObj)
-                        self.inputParentDict[key] = parentObj
-                else:
-                    self.inputParentDict[key] = inputObj
-                inputs.append(attr)
-
-            outputs = [attr for _, attr in
-                       self.protocol.iterOutputAttributes()]
-            self.outputStr = pwobj.String(Message.LABEL_OUTPUT)
-            objs = inputParents + inputs + [self.outputStr] + outputs
-        return objs
-
-    def _visualizeObject(self, ViewerClass, obj):
-        viewer = ViewerClass(project=self.protocol.getProject(),
-                             protocol=self.protocol,
-                             parent=self.parent.windows)
-        viewer.visualize(obj, windows=self.parent.windows)
-
-    def _editObject(self, obj):
-        """Open the Edit GUI Form given an instance"""
-        pwgui.dialog.EditObjectDialog(self.parent, Message.TITLE_EDIT_OBJECT,
-                                      obj, self.mapper)
-
-    def _deleteObject(self, obj):
-        """ Remove unnecessary output, specially for Coordinates. """
-        prot = self.protocol
-        try:
-            objLabel = self.getObjectLabel(obj, prot)
-            if self.parent.windows.askYesNo("Delete object",
-                                            "Are you sure to delete *%s* object?"
-                                            % objLabel):
-                prot.getProject().deleteProtocolOutput(prot, obj)
-                self.parent._fillSummary()
-                self.parent.windows.showInfo("Object *%s* successfully deleted."
-                                             % objLabel)
-        except Exception as ex:
-            self.parent.windows.showError(str(ex))
-
-    @staticmethod
-    def getObjectPreview(obj):
-        desc = "<name>: " + obj.getName()
-        return None, desc
-
-    def getObjectActions(self, obj):
-        if isinstance(obj, pwobj.Pointer):
-            obj = obj.get()
-            isPointer = True
-        else:
-            isPointer = False
-        actions = []
-
-        viewers = Config.getDomain().findViewers(obj.getClassName(), DESKTOP_TKINTER)
-
-        def viewerCallback(viewer):
-            return lambda: self._visualizeObject(viewer, obj)
-
-        for v in viewers:
-            actions.append(('Open with %s' % v.__name__,
-                            viewerCallback(v),
-                            Icon.ACTION_VISUALIZE))
-        # EDIT
-        actions.append((Message.LABEL_EDIT,
-                        lambda: self._editObject(obj),
-                        Icon.ACTION_EDIT))
-        # DELETE
-        # Special case to allow delete outputCoordinates
-        # since we can end up with several outputs and
-        # we may want to clean up
-        if self.protocol.allowsDelete(obj) and not isPointer:
-            actions.append((Message.LABEL_DELETE_ACTION,
-                            lambda: self._deleteObject(obj),
-                            Icon.ACTION_DELETE))
-        return actions
-
-    @staticmethod
-    def getObjectLabel(obj, parent):
-        """ We will try to show in the list the string representation
-        that is more readable for the user to pick the desired object.
-        """
-        label = 'None'
-        if obj:
-            label = obj.getObjLabel()
-            if not len(label.strip()):
-                parentLabel = parent.getObjLabel() if parent else 'None'
-                label = "%s -> %s" % (parentLabel, obj.getLastName())
-        return label
-
-    def getObjectInfo(self, obj):
-        if obj is None or not obj.hasValue():
-            return None
-
-        if isinstance(obj, pwobj.String):
-            value = obj.get()
-            info = {'key': value, 'text': value, 'values': '', 'open': True}
-            if hasattr(obj, '_parentKey'):
-                info['parent'] = self.inputParentDict[obj._parentKey]
-        else:
-            # All attributes are considered output, unless they are pointers
-            image = Icon.ACTION_OUT
-            parent = self.outputStr
-
-            if isinstance(obj, pwobj.Pointer):
-                name = obj.getLastName()
-                # Remove ugly item notations inside lists
-                name = name.replace('__item__000', '')
-                # Consider Pointer as inputs
-                image = getattr(obj, '_icon', '')
-                parent = self.inputParentDict[obj._parentKey]
-
-                suffix = ''
-                if obj.hasExtended():
-                    # getExtended method remove old attributes conventions.
-                    extendedValue = obj.getExtended()
-                    if obj.hasExtended():
-                        suffix = '[%s]' % extendedValue
-                    # else:
-                    #     suffix = '[Item %s]' % extendedValue
-
-                    # Tolerate loading projects:
-                    # When having only the project sqlite..an obj.get() will
-                    # the load of the set...and if it is missing this whole
-                    # "thread" fails.
-                    try:
-                        labelObj = obj.get()
-                        if labelObj is None:
-                            labelObj = obj.getObjValue()
-                            suffix = ''
-
-                    except Exception as e:
-                        return {'parent': parent, 'image': image, 'text': name,
-                                'values': ("Couldn't read object attributes.",)}
-                else:
-                    labelObj = obj.get()
-
-                objKey = obj._parentKey + str(labelObj.getObjId())
-                label = self.getObjectLabel(labelObj,
-                                            self.mapper.getParent(labelObj))
-                name += '   (from %s %s)' % (label, suffix)
-            else:
-                name = self.getObjectLabel(obj, self.protocol)
-                objKey = str(obj.getObjId())
-                labelObj = obj
-
-            # To tolerate str(labelObj) in case xmippLib is missing, but
-            # still being able to open a project.
-            try:
-                value = str(labelObj)
-            except Exception as e:
-                print("Can not convert object %s - %s to string." % (objKey, name))
-                value = str(e)
-
-            info = {'key': objKey, 'parent': parent, 'image': image,
-                    'text': name, 'values': (value,)}
-        return info
 
 
 # noinspection PyAttributeOutsideInit
@@ -649,6 +100,7 @@ class ProtocolsView(tk.Frame):
         self.root.bind("<Control-a>", self._selectAllProtocols)
         self.root.bind("<Control-t>", self._toggleColorScheme)
         self.root.bind("<Control-d>", self._toggleDebug)
+        self.root.bind("<Control-l>", self._locateProtocol)
         self.root.bind("<F2>", self._F2Pressed)
 
         if Config.debugOn():
@@ -822,6 +274,13 @@ class ProtocolsView(tk.Frame):
 
         self._updateSelection()
 
+        # Move to the selected protocol
+        if self._isSingleSelection():
+            prot = self.getSelectedProtocol()
+            node = self.runsGraph.getNode(str(prot.getObjId()))
+            self._selectNode(node)
+
+
         # Add all tabs
 
         tab.add(dframe, text=Message.LABEL_SUMMARY)
@@ -962,6 +421,33 @@ class ProtocolsView(tk.Frame):
         window = SearchProtocolWindow(self.windows)
         window.show()
 
+    def _locateProtocol(self, e=None):
+
+        window = SearchRunWindow(self.windows, self.runsGraph, onDoubleClick=self._onRunClick)
+        window.show()
+        # self._moveCanvas(0,1)
+
+    def _onRunClick(self, e=None):
+        """ Callback to be called when a click happens o a run in the SearchRunWindow.tree"""
+        tree = e.widget
+        protId = tree.getFirst()
+        node = self.runsGraph.getNode(protId)
+        self._selectNode(node)
+
+    def _selectNode(self, node):
+
+        x = node.x
+        y = node.y
+        self._moveCanvas(x,y)
+
+        # Select the protocol
+        self._selectItemProtocol(node.run)
+        self.refreshDisplayedRuns()
+
+    def _moveCanvas(self, X, Y):
+
+        self.runsGraphCanvas.moveTo(X, Y)
+
     def createActionToolbar(self):
         """ Prepare the buttons that will be available for protocol actions. """
 
@@ -970,7 +456,7 @@ class ProtocolsView(tk.Frame):
                            ACTION_STEPS, ACTION_BROWSE, ACTION_DB,
                            ACTION_STOP, ACTION_CONTINUE, ACTION_RESULTS,
                            ACTION_EXPORT, ACTION_EXPORT_UPLOAD, ACTION_COLLAPSE,
-                           ACTION_EXPAND, ACTION_LABELS]
+                           ACTION_EXPAND, ACTION_LABELS, ACTION_SEARCH]
 
         def addButton(action, text, toolbar):
             btn = tk.Label(toolbar, text=text,
@@ -2381,6 +1867,8 @@ class ProtocolsView(tk.Frame):
                     self._stopWorkFlow(action)
                 elif action == ACTION_RESET_WORKFLOW:
                     self._resetWorkFlow(action)
+                elif action == ACTION_SEARCH:
+                    self._searchProtocol()
 
             except Exception as ex:
                 self.windows.showError(str(ex))
@@ -2436,511 +1924,3 @@ class RunBox(pwgui.TextBox):
         self.nodeInfo.setPosition(self.x, self.y)
 
 
-def inspectObj(obj, filename, prefix='', maxDeep=5, inspectDetail=2, memoryDict=None):
-    """ Creates a .CSV file in the filename path with
-        all its members and recursively with a certain maxDeep,
-        if maxDeep=0 means no maxDeep (until all members are inspected).
-        
-        inspectDetail can be:
-         - 1: All attributes are shown
-         - 2: All attributes are shown and iterable values are also inspected
-
-        prefix and memoryDict will be updated in the recursive entries:
-         - prefix is a compound of the two first columns (DEEP and Tree)
-         - memoryDict is a dictionary with the memory address and an identifier
-    """
-    END_LINE = '\n'  # end of line char
-    COL_DELIM = '\t'  # column delimiter
-    INDENT_COUNTER = '/'  # character append in each indention (it's not written)
-
-    NEW_CHILD = '  |------>  '  # new item indention
-    BAR_CHILD = '  | ' + INDENT_COUNTER  # bar indention
-    END_CHILD = ('       -- '+COL_DELIM)*4 + END_LINE  # Child ending
-    column1 = '    - Name - ' + COL_DELIM
-    column2 = '    - Type - ' + COL_DELIM
-    column3 = '    - Value - ' + COL_DELIM
-    column4 = '  - Memory Address -'
-
-    #  Constants to distinguish the first, last and middle rows
-    IS_FIRST = 1
-    IS_LAST = -1
-    IS_MIDDLE = 0
-
-    memoryDict = memoryDict or {}
-
-    def writeRow(name, value, prefix, posList=False):
-        """ Writes a row item. """
-        # we will avoid to recursively print the items wrote before 
-        #  (ie. with the same memory address), thus we store a dict with the
-        #  addresses and the flag isNew is properly set
-        if str(hex(id(value))) in memoryDict:
-            memorySTR = memoryDict[str(hex(id(value)))]
-            isNew = False
-        else:
-            # if the item is new, we save its memory address in the memoryDict
-            #   and we pass the name and the line on the file as a reference.
-            memorySTR = str(hex(id(value)))
-            file = open(filename, 'r')
-            lineNum = str(len(file.readlines())+1)
-            file.close()
-            nameDict = str(name)[0:15]+' ...' if len(str(name)) > 25 else str(name)
-            memoryDict[str(hex(id(value)))] = '>>> '+nameDict + ' - L:'+lineNum
-            isNew = True
-        
-        if posList:
-            # if we have a List, the third column is 'pos/lenght'
-            thirdCol = posList
-        else:
-            # else, we print the value avoiding the EndOfLine char (// instead)
-            thirdCol = str(value).replace(END_LINE, ' // ')
-
-        # we will print the indentation deep number in the first row
-        indentionDeep = prefix.count(INDENT_COUNTER)
-        deepStr = str(indentionDeep) + COL_DELIM
-
-        # the prefix without the indentCounters is 
-        #   the tree to be printed in the 2nd row
-        prefixToWrite = prefix.replace(INDENT_COUNTER, '')
-        
-        file = open(filename, 'a')   
-        file.write(deepStr + prefixToWrite + COL_DELIM +
-                   str(name) + COL_DELIM +
-                   str(type(value)) + COL_DELIM +
-                   thirdCol + COL_DELIM +
-                   memorySTR + END_LINE)
-        file.close()
-
-        return isNew
-
-    def recursivePrint(value, prefix, isFirstOrLast):
-        """ We print the childs items of tuples, lists, dicts and classes. """ 
-
-        # if it's the last item, its childs has not the bar indention
-        if isFirstOrLast == IS_LAST:  # void indention when no more items
-            prefixList = prefix.split(INDENT_COUNTER)
-            prefixList[-2] = prefixList[-2].replace('|', ' ')
-            prefix = INDENT_COUNTER.join(prefixList)
-
-        # recursive step with the new prefix and memory dict.
-        inspectObj(value, filename, prefix+BAR_CHILD, maxDeep, inspectDetail, 
-                   memoryDict)
-        
-        if isFirstOrLast == IS_FIRST:
-            deepStr = str(indentionDeep) + COL_DELIM
-        else:
-            # When it was not the first item, the deep is increased
-            #   to improve the readability when filter 
-            deepStr = str(indentionDeep+1) + COL_DELIM
-
-        prefix = prefix.replace(INDENT_COUNTER, '') + COL_DELIM
-
-        # We introduce the end of the child and 
-        #   also the next header while it is not the last
-        file = open(filename, 'a')
-        file.write(deepStr + prefix + END_CHILD)
-        if isFirstOrLast != IS_LAST:
-            # header
-            file.write(deepStr + prefix + 
-                       column1 + column2 + column3 + column4 + END_LINE)
-        file.close()
-
-    def isIterable(obj):
-        """ Returns true if obj is a tuple, list, dict or calls. """
-        isTupleListDict = (isinstance(obj, tuple) or
-                           isinstance(obj, dict) or
-                           isinstance(obj, list)) and len(value) > 1
-
-        # FIX ME: I don't know how to assert if is a class or not... 
-        isClass = str(type(obj))[1] == 'c'
-
-        return isClass or (isTupleListDict and inspectDetail < 2)
-
-    indentionDeep = prefix.count(INDENT_COUNTER)
-    if indentionDeep == 0:
-        prefix = ' - Root - '
-
-        # dict with name and value pairs of the members
-        if len(obj) == 1:
-            # if only one obj is passed in the input list,
-            #   we directly inspect that obj.
-            obj_dict = obj[0].__dict__
-            obj = obj[0]
-
-        #  setting the header row
-        treeHeader = ' - Print on ' + str(dt.datetime.now())
-        prefixHeader = '-DEEP-' + COL_DELIM + treeHeader + COL_DELIM
-        col1 = '    - Name - (value for Lists and Tuples)' + COL_DELIM
-        col3 = '    - Value - (Pos./Len for Lists and Tuples) ' + COL_DELIM
-
-        #  writing the header row
-        file = open(filename, 'w')
-        file.write(prefixHeader + col1 + column2 + col3 + column4 + END_LINE)
-        file.close()
-
-        #  writing the root object
-        writeRow(obj.__class__.__name__, obj, prefix)
-        #  adding the child bar to the prefix
-        prefix = '  ' + BAR_CHILD
-    else:
-        # firsts settings depending on the type of the obj
-        if str(type(obj))[1] == 'c':
-            obj_dict = obj.__dict__
-        elif (isinstance(obj, tuple) or
-              isinstance(obj, list)):
-            column1 = '    - Value - ' + COL_DELIM
-            column3 = '  - Pos./Len. - ' + COL_DELIM
-        elif isinstance(obj, dict):
-            column1 = '    - Key - ' + COL_DELIM
-            obj_dict = obj
-        else:  # if is not of the type above it not make sense to continue
-            return
-
-    indentionDeep = prefix.count(INDENT_COUNTER)
-    deepStr = str(indentionDeep) + COL_DELIM
-    isBelowMaxDeep = indentionDeep < maxDeep if maxDeep > 0 else True 
-
-    prefixToWrite = prefix.replace(INDENT_COUNTER, '') + COL_DELIM
-    file = open(filename, 'a')
-    file.write(deepStr + prefixToWrite + 
-               column1 + column2 + column3 + column4 + END_LINE)
-    file.close()
-
-    #  we update the prefix to put the NEW_CHILD string  ( |----> )
-    prefixList = prefix.split(INDENT_COUNTER)
-    prefixList[-2] = NEW_CHILD
-    #  we return to the string structure
-    #    with a certain indention if it's the root
-    prefixToWrite = '  ' + INDENT_COUNTER.join(prefixList) if indentionDeep == 1 \
-        else INDENT_COUNTER.join(prefixList)
-
-    isNew = True
-    if str(type(obj))[1] == 'c' or isinstance(obj, dict):
-        counter = 0
-        for key, value in obj_dict.items():
-            counter += 1
-            # write the variable
-            isNew = writeRow(key, value, prefixToWrite)
-
-            # managing the extremes of the loop
-            if counter == 1:
-                isFirstOrLast = IS_FIRST
-            elif counter == len(obj_dict):
-                isFirstOrLast = IS_LAST
-            else:
-                isFirstOrLast = IS_MIDDLE
-
-            # show attributes for objects and items for lists and tuples 
-            if isBelowMaxDeep and isNew and isIterable(value):
-                recursivePrint(value, prefix, isFirstOrLast)
-    else:
-        for i in range(0, len(obj)):
-            # write the variable
-            isNew = writeRow(obj[i], obj[i], prefixToWrite, 
-                             str(i+1)+'/'+str(len(obj)))
-
-            # managing the extremes of the loop
-            if i == 0:
-                isFirstOrLast = IS_FIRST
-            elif len(obj) == i+1:
-                isFirstOrLast = IS_LAST
-            else:
-                isFirstOrLast = IS_MIDDLE
-
-            # show attributes for objects and items for lists and tuples 
-            if isBelowMaxDeep and isNew and isIterable(obj[i]):
-                recursivePrint(obj[i], prefix, isFirstOrLast)
-
-
-class ProtocolTreeConfig:
-    """ Handler class that groups functions and constants
-    related to the protocols tree configuration.
-    """
-    ALL_PROTOCOLS = "All"
-    TAG_PROTOCOL_DISABLED = 'protocol-disabled'
-    TAG_PROTOCOL = 'protocol'
-    TAG_SECTION = 'section'
-    TAG_PROTOCOL_GROUP = 'protocol_group'
-    TAG_PROTOCOL_BETA = 'protocol_beta'
-    TAG_PROTOCOL_NEW = 'protocol_new'
-    PLUGIN_CONFIG_PROTOCOLS = 'protocols.conf'
-
-    @classmethod
-    def getProtocolTag(cls, isInstalled, isBeta=False, isNew=False):
-        """ Return the proper tag depending if the protocol is installed or not.
-        """
-        if isInstalled:
-            if isBeta:
-                return cls.TAG_PROTOCOL_BETA
-            elif isNew:
-                return cls.TAG_PROTOCOL_NEW
-            return cls.TAG_PROTOCOL
-        else:
-            return cls.TAG_PROTOCOL_DISABLED
-
-    @classmethod
-    def isAFinalProtocol(cls, v, k):
-        if (issubclass(v, ProtocolViewer) or
-                v.isBase() or v.isDisabled()):
-            return False
-
-        return v.__name__ == k
-
-    @classmethod
-    def __addToTree(cls, menu, item, checkFunction=None):
-        """ Helper function to recursively add items to a menu.
-        Add item (a dictionary that can contain more dictionaries) to menu
-        If check function is added will use it to check if the value must be added.
-        """
-        children = item.pop('children', [])
-
-        if checkFunction is not None:
-            add = checkFunction(item)
-            if not add:
-                return
-        subMenu = menu.addSubMenu(**item)  # we expect item={'text': ...}
-        for child in children:
-            cls.__addToTree(subMenu, child, checkFunction)  # add recursively to sub-menu
-
-        return subMenu
-
-    @classmethod
-    def __inSubMenu(cls, child, subMenu):
-        """
-        Return True if child belongs to subMenu
-        """
-        for ch in subMenu:
-            if cls.__isProtocol(child):
-                if ch.value is not None and ch.value == child['value']:
-                    return ch
-            elif ch.text == child['text']:
-                return ch
-        return None
-
-    @classmethod
-    def _orderSubMenu(cls, session):
-        """
-        Order all children of a given session:
-        The protocols first, then the sessions(the 'more' session at the end)
-        """
-        lengthSession = len(session.childs)
-        if lengthSession > 1:
-            childs = session.childs
-            lastChildPos = lengthSession - 1
-            if childs[lastChildPos].tag == cls.TAG_PROTOCOL:
-                for i in range(lastChildPos - 1, -1, -1):
-                    if childs[i].tag == cls.TAG_PROTOCOL:
-                        break
-                    else:
-                        tmp = childs[i + 1]
-                        childs[i + 1] = childs[i]
-                        childs[i] = tmp
-            else:
-                for i in range(lastChildPos - 1, -1, -1):
-                    if childs[i].tag == cls.TAG_PROTOCOL:
-                        break
-                    elif 'more' in str(childs[i].text).lower():
-                        tmp = childs[i + 1]
-                        childs[i + 1] = childs[i]
-                        childs[i] = tmp
-
-    @classmethod
-    def __findTreeLocation(cls, subMenu, children, parent):
-        """
-        Locate the protocol position in the given view
-        """
-        for child in children:
-            sm = cls.__inSubMenu(child, subMenu)
-            if sm is None:
-                cls.__addToTree(parent, child, cls.__checkItem)
-                cls._orderSubMenu(parent)
-            elif child['tag'] == cls.TAG_PROTOCOL_GROUP or child['tag'] == cls.TAG_SECTION:
-                cls.__findTreeLocation(sm.childs, child['children'], sm)
-    @classmethod
-    def __isProtocol(cls, dict):
-        """ True inf the item has a key named tag with protocol as value"""
-        return dict["tag"] == cls.TAG_PROTOCOL
-
-    @classmethod
-    def __isProtocolNode(cls, node):
-        """ True if tag attribute is protocol"""
-        return node.tag == cls.TAG_PROTOCOL
-
-
-    @classmethod
-    def __checkItem(cls, item):
-        """ Function to check if the protocol has to be added or not.
-        Params:
-            item: {"tag": "protocol", "value": "ProtImportMovies",
-                   "text": "import movies"}
-        """
-        if not cls.__isProtocol(item):
-            return True
-
-        # It is a protocol as this point, get the class name and
-        # check if it is disabled
-        protClassName = item["value"]
-        protClass = Config.getDomain().getProtocols().get(protClassName)
-        icon = 'python_file.gif'
-        if protClass is not None:
-            if protClass.isBeta():
-                icon = "beta.gif"
-            elif protClass.isNew():
-                icon = "new.gif"
-        item['icon'] = icon
-        return False if protClass is None else not protClass.isDisabled()
-
-    @classmethod
-    def __addAllProtocols(cls, domain, protocols):
-        # Add all protocols
-        allProts = domain.getProtocols()
-
-        # Sort the dictionary
-        allProtsSorted = OrderedDict(sorted(allProts.items(),
-                                            key=lambda e: e[1].getClassLabel()))
-        allProtMenu = ProtocolConfig(cls.ALL_PROTOCOLS)
-        packages = {}
-
-        # Group protocols by package name
-        for k, v in allProtsSorted.items():
-            if cls.isAFinalProtocol(v, k):
-                packageName = v.getClassPackageName()
-                # Get the package submenu
-                packageMenu = packages.get(packageName)
-
-                # If no package menu available
-                if packageMenu is None:
-                    # Add it to the menu ...
-                    packageLine = {"tag": "package", "value": packageName,
-                                   "text": packageName}
-                    packageMenu = cls.__addToTree(allProtMenu, packageLine)
-
-                    # Store it in the dict
-                    packages[packageName] = packageMenu
-
-                # Add the protocol
-                tag = cls.getProtocolTag(v.isInstalled(), v.isBeta(), v.isNew())
-
-                protLine = {"tag": tag, "value": k,
-                            "text": v.getClassLabel(prependPackageName=False)}
-
-                cls.__addToTree(packageMenu, protLine)
-
-        protocols[cls.ALL_PROTOCOLS] = allProtMenu
-
-    @classmethod
-    def __addProtocolsFromConf(cls, protocols, protocolsConfPath):
-        """
-        Load the protocols in the tree from a given protocols.conf file,
-        either the global one in Scipion or defined in a plugin.
-        """
-        # Populate the protocols menu from the plugin config file.
-        if os.path.exists(protocolsConfPath):
-            cp = ConfigParser()
-            cp.optionxform = str  # keep case
-            cp.read(protocolsConfPath)
-            #  Ensure that the protocols section exists
-            if cp.has_section('PROTOCOLS'):
-                for menuName in cp.options('PROTOCOLS'):
-                    if menuName not in protocols:  # The view has not been inserted
-                        menu = ProtocolConfig(menuName)
-                        children = json.loads(cp.get('PROTOCOLS', menuName))
-                        for child in children:
-                            cls.__addToTree(menu, child, cls.__checkItem)
-                        protocols[menuName] = menu
-                    else:  # The view has been inserted
-                        menu = protocols.get(menuName)
-                        children = json.loads(cp.get('PROTOCOLS',
-                                                     menuName))
-                        cls.__findTreeLocation(menu.childs, children, menu)
-
-    @classmethod
-    def load(cls, domain, protocolsConf):
-        """ Read the protocol configuration from a .conf file similar to the
-        one in scipion/config/protocols.conf,
-        which is the default one when no file is passed.
-        """
-        protocols = dict()
-        # Read the protocols.conf from Scipion (base) and create an initial
-        # tree view
-        cls.__addProtocolsFromConf(protocols, protocolsConf)
-
-        # Read the protocols.conf of any installed plugin
-        pluginDict = domain.getPlugins()
-        pluginList = cls.__orderByPriority(pluginDict.keys(),
-                                           priorityPluginList=Config.getPriorityPackageList())
-        for pluginName in pluginList:
-            try:
-
-                # if the plugin has a path
-                if pwutils.isModuleLoaded(pluginName) and pwutils.isModuleAFolder(pluginName):
-                    # Locate the plugin protocols.conf file
-                    protocolsConfPath = os.path.join(
-                        pluginDict[pluginName].__path__[0],
-                        cls.PLUGIN_CONFIG_PROTOCOLS)
-                    cls.__addProtocolsFromConf(protocols, protocolsConfPath)
-
-            except Exception as e:
-                print('Failed to read settings. The reported error was:\n  %s\n'
-                      'To solve it, fix %s and run again.' % (
-                          e, os.path.abspath(protocolsConfPath)))
-
-        # Clean empty sections
-        cls._hideEmptySections(protocols)
-
-        # Add all protocols to All view
-        cls.__addAllProtocols(domain, protocols)
-
-        return protocols
-
-    @classmethod
-    def _hideEmptySections(cls, protocols):
-        """ Cleans all empty sections in the tree"""
-
-        for protConf in protocols.values():
-            cls._setVisibility(protConf)
-
-    @classmethod
-    def _setVisibility(cls, node):
-        """ Sets the visibility of a node based on the presence of a leaf hanging form it"""
-        if cls.__isProtocolNode(node):
-            # Default visibility value is true. No need to set it again
-            return True
-
-        anyLeaf = False
-
-        for child in node.childs:
-            # NOTE: since python short circuits this, _setVisibility must be called always. So not swap!!
-            anyLeaf = cls._setVisibility(child) or anyLeaf
-
-        node.visible = anyLeaf
-
-        return anyLeaf
-    @classmethod
-    def __orderByPriority(cls, pluginList, priorityPluginList):
-        if priorityPluginList:
-            sortedPluginList = priorityPluginList + [pluginName for pluginName in pluginList
-                                                     if pluginName not in priorityPluginList]
-            return sortedPluginList
-        else:
-            return pluginList
-
-
-class ProtocolConfig(MenuConfig):
-    """Store protocols configuration """
-
-    def __init__(self, text=None, value=None, **args):
-        MenuConfig.__init__(self, text, value, **args)
-        if 'openItem' not in args:
-            self.openItem = self.tag != 'protocol_base'
-
-    def addSubMenu(self, text, value=None, shortCut=None, **args):
-        if 'icon' not in args:
-            tag = args.get('tag', None)
-            if tag == 'protocol_base':
-                args['icon'] = 'class_obj.gif'
-
-        args['shortCut'] = shortCut
-        return MenuConfig.addSubMenu(self, text, value, **args)
-
-    def __str__(self):
-        return self.text
